UI: Fix issues related to multi-track stream output
The audio track was accidentally being set as the index of the audio
encoder. If the user picked track 2 in the OBS settings this would
cause the multi-track RTMP output code to activate, resulting in
non-obvious failed streams due to lack of server support (Twitch for
example would just stop reading from the network, making it look like
a network / connectivity failure).