UI: Check explicitly if outputs needs initializing
With the addition of Virtual Camera, an active output no longer
guarantees that all of the outputs have been set up. Instead of checking
for an active output, check if the OBSEncoder fields need setting up.

Fixes crashing and other errors if Virtual Camera was activated followed
by a stream / record attempt.