libobs: Optimize memory handling functions
malloc(0) is bad, so don't bother going out of our way to pretend to
handle it. Set the base_allocator to a const struct, so that inline
optimizations can replace calls to bmalloc with direct CRT calls. This
reduces the amount of instructions required per inlined bmalloc / bfree
call.

This turns the base_set_allocator call into a no-op, but it doesn't seem
like anything ever called it anyway. Possible risk of crashes in 3rd
party plugins if they call bmalloc(0) but that is bad behavior to begin
with.