libobs: Optimize Windows timer functions
os_gettime_ns is one of the hottest functions in libobs, as it's called
in a spin loop for timing of media sources and the graphics thread.
Switching to integer multiplication results in a large reduction in
micro-ops and latency on all CPUs. This doesn't translate into a real
world performance gain since we're still spinning in a loop, but it
might increase timing accuracy or reduce power consumption.

Note that as of #5114 the impact is less severe as media sources no
longer hit this function.