libobs: Implement OBS_IMPRECISE_FLOAT macros
These allow relaxing some of the FPU math constraints on a per-function
basis, allowing the compiler to do auto-vectorization of loops that
previously weren't vectorizable due to associative math rules. This is
used in a couple of places in the audio pipeline where the potential
loss of precision is insignificant.

Currently only implemented for MSVC and GCC as this isn't supported by
all compilers.