UI: Try to remove local MSVC redists on startup
If something has copied MSVC redists to the OBS install directory, the
user will be stuck in a loop of "Outdated redists" as the  local copies
take precedence over the system ones. Using the same technique as the
updater, we can move the in-use files away and restart OBS to resolve
this.